# src/main/java/com/nelumbo/zoo_api/dto/SpeciesRequest.java
package com.nelumbo.zoo_api.dto;

import com.nelumbo.zoo_api.validation.annotations.UniqueSpeciesName;
import jakarta.validation.constraints.NotBlank;
import org.springframework.validation.annotation.Validated;

@Validated
public record SpeciesRequest(@NotBlank @UniqueSpeciesName String name) {}