# src/main/resources/static/docs/openapi.yaml
openapi: "3.0.0"
info:
  title: ZOO System
  description: API para gestión de zoologicos. Desarrollada con Spring boot y PostgreSQL.
  contact:
    name: Brayan Barco
    url: alex.barco.maicol@gmail.com
    email: alex.barco.maicol@gmail.com
  version: 1.0.0
  x-products:
    - id: zoo-api
      name: ZOO System
      description: API de gestión de Zoologicos
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
  - url: http://localhost:8080
    description: Servidor de desarrollo
  - url: http://localhost:8080
    description: Servidor de QA
  - url: http://localhost:8080
    description: Servidor de producción
security:
  - Security_Token: []
paths:
...
  /api/zones/{id}:
    get:
      tags:
        - zone-controller
      summary: Listar zonas
      description: Obtener todas las zonas disponibles
      operationId: getZoneById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SuccessResponseDTOZoneResponse'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
    put:
      tags:
        - zone-controller
      summary: Actualizar zona
      description: Actualizar información de zona (solo Admin)
      operationId: updateZone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZoneRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SuccessResponseDTOZoneResponse'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
    delete:
      tags:
        - zone-controller
      summary: Eliminar zona
      description: Eliminar zona (solo Admin, si no tiene animales)
      operationId: deleteZone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SuccessResponseDTOVoid'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
  /api/species/{id}:
    get:
      tags:
        - species-controller
      summary: Listar especies
      description: Obtener todas las especies registradas
      operationId: getSpeciesById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SuccessResponseDTOSpeciesResponse'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
    put:
      tags:
        - species-controller
      summary: Actualizar especie
      description: Actualizar información de especie (solo Admin)
      operationId: updateSpecies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeciesRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SuccessResponseDTOSpeciesResponse'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
    delete:
      tags:
        - species-controller
      summary: Eliminar especie
      description: Eliminar especie (solo Admin, si no tiene animales)
      operationId: deleteSpecies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SuccessResponseDTOVoid'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
...
  /api/zones:
    get:
      tags:
        - zone-controller
      summary: Listar zonas
      description: Obtener todas las zonas disponibles
      operationId: getAllZones
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SuccessResponseDTOListZoneResponse'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
    post:
      tags:
        - zone-controller
      summary: Crear zona
      description: Crear nueva zona (solo Admin)
      operationId: createZone
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZoneRequest'
        required: true
      responses:
        '201':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SuccessResponseDTOZoneResponse'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
  /api/species:
    get:
      tags:
        - species-controller
      summary: Listar especies
      description: Obtener todas las especies registradas
      operationId: getAllSpecies
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SuccessResponseDTOListSpeciesResponse'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
    post:
      tags:
        - species-controller
      summary: Crear especie
      description: Registrar nueva especie (solo Admin)
      operationId: createSpecies
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeciesRequest'
        required: true
      responses:
        '201':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SuccessResponseDTOSpeciesResponse'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
...
components:
  schemas:
    StandardError:
      type: object
      properties:
        code:
          type: string
          minLength: 2
          maxLength: 20
          description: Código del error
          example: PERR100
        description:
          type: string
          minLength: 2
          maxLength: 250
          description: Descripción del error
          example: Error de validación
        field:
          type: string
          minLength: 2
          maxLength: 100
          description: Campo con error
          example: email
        traceId:
          type: string
          format: uuid
          description: ID de traza para diagnóstico
          example: "550e8400-e29b-41d4-a716-446655440000"
      required:
        - code
        - description
    StandardErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/StandardError'
    ZoneRequest:
      type: object
      properties:
        name:
          type: string
    AnimalSimpleResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    SuccessResponseDTOZoneResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ZoneResponse'
        errors: {}
    ZoneResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        animals:
          type: array
          items:
            $ref: '#/components/schemas/AnimalSimpleResponse'
    SpeciesRequest:
      type: object
      properties:
        name:
          type: string
    SpeciesResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    SuccessResponseDTOSpeciesResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/SpeciesResponse'
        errors: {}
...
    SuccessResponseDTOListZoneResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ZoneResponse'
        errors: {}
...
    SuccessResponseDTOListSpeciesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SpeciesResponse'
        errors: {}
...
    SuccessResponseDTOVoid:
      type: object
      properties:
        data: {}
        errors: {}
  securitySchemes:
    Security_Token:
      type: http
      description: Access Token For My API
      scheme: bearer
      bearerFormat: JWT
  responses:
    badRequestResponse:
      description: Solicitud incorrecta
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            BadRequest:
              value:
                errors:
                  - code: "400"
                    description: "El campo ' ' es requerido"
                    field: "email"
                    traceId: "550e8400-e29b-41d4-a716-446655440000"
    unauthorizedResponse:
      description: No autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            Unauthorized:
              value:
                errors:
                  - code: "401"
                    description: "Token de acceso no válido o expirado"
    forbiddenResponse:
      description: Prohibido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            Forbidden:
              value:
                errors:
                  - code: "403"
                    description: "No tiene permisos para acceder a este recurso"
    notFoundResponse:
      description: No encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            NotFound:
              value:
                errors:
                  - code: "404"
                    description: "Elemento no encontrado"
    internalServerErrorResponse:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            InternalError:
              value:
                errors:
                  - code: "500"
                    description: "Error interno del servidor"
    notAcceptableResponse:
      description: No aceptable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            NotAcceptable:
              value:
                errors:
                  - code: "406"
                    description: "El encabezado Accept no incluye un tipo de medio soportado"
    payloadTooLargeResponse:
      description: Carga útil demasiado grande
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            PayloadTooLarge:
              value:
                errors:
                  - code: "413"
                    description: "La carga útil de la solicitud excede el límite permitido"
    unsupportedMediaTypeResponse:
      description: Tipo de medio no soportado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            UnsupportedMediaType:
              value:
                errors:
                  - code: "415"
                    description: "El tipo de medio de la solicitud no está soportado"
    tooManyRequestsResponse:
      description: Demasiadas solicitudes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            TooManyRequests:
              value:
                errors:
                  - code: "429"
                    description: "Se ha excedido el límite de tasa de solicitudes"
    serviceUnavailableResponse:
      description: Servicio no disponible
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            ServiceUnavailable:
              value:
                errors:
                  - code: "503"
                    description: "El servicio no está disponible temporalmente"